Separate model from internal
