+++ internal/android/nonce_test.go
@@ -17,7 +17,7 @@ package android
 import (
 	"testing"
 
-	"github.com/google/exposure-notifications-server/internal/model"
+	"github.com/google/exposure-notifications-server/model"
 )
 
 // Data from this test was generated by the Android reference application.
