+++ internal/android/safetynet_test.go
@@ -20,7 +20,7 @@ import (
 	"time"
 
 	"github.com/google/exposure-notifications-server/internal/base64util"
-	"github.com/google/exposure-notifications-server/internal/model"
+	"github.com/google/exposure-notifications-server/model"
 )
 
 const (
