+++ internal/api/jsonutil/unmarshal_test.go
@@ -23,7 +23,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/google/exposure-notifications-server/internal/model"
+	"github.com/google/exposure-notifications-server/model"
 
 	"github.com/google/go-cmp/cmp"
 )
