+++ internal/database/federationin.go
@@ -19,7 +19,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/google/exposure-notifications-server/internal/model"
+	"github.com/google/exposure-notifications-server/model"
 
 	pgx "github.com/jackc/pgx/v4"
 )
