+++ model/exposure.go
