+++ tools/export-config/main.go
@@ -23,7 +23,7 @@ import (
 	"time"
 
 	"github.com/google/exposure-notifications-server/internal/database"
-	"github.com/google/exposure-notifications-server/internal/model"
+	"github.com/google/exposure-notifications-server/model"
 	"github.com/kelseyhightower/envconfig"
 )
